requirements.txt
pytest~=7.4.0
allure-pytest
selene>=2.0.0rc4
Appium-Python-Client~=3.0.0
pydantic~=2.3.0
requests~=2.31.0
python-dotenv~=1.0.0